# openapi.yaml
openapi: 3.0.3
info:
  title: IRIA Ecosystem Platform API (Hackathon MVP)
  version: 1.0.0
  description: |
    API contract for the Azerbaijan Startup Ecosystem Platform (IRIA Hackathon MVP).

    Platform goals:
    - Startup registry and ecosystem stakeholder management
    - AI-assisted startup analysis (SWOT, success score, market insights)
    - Similar startup discovery
    - Program application workflows
    - Investor interest tracking
    - KPI dashboards for IRIA/Government
    - Role-based access control and audit visibility
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.iria-ecosystem.az/api/v1
    description: Production (planned)
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      description: |
        Creates a new user account. A user can be one of:
        startup, investor, mentor, iria_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              full_name: "Aysel Mammadova"
              email: "aysel@example.com"
              password: "StrongPass123!"
              role: "startup"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "aysel@example.com"
              password: "StrongPass123!"
      responses:
        "200":
          description: Login successful (JWT issued)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /events:
    get:
      tags: [Events]
      summary: List ecosystem events and hackathons
      security: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [hackathon, accelerator, incubator, meetup, grant]
        - in: query
          name: sector
          schema:
            type: string
        - in: query
          name: topic
          schema:
            type: string
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"

    post:
      tags: [Events]
      summary: Create event/program listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreateRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request payload or query params
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    UserRole:
      type: string
      enum: [startup, investor, mentor, iria_admin]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
          required: [code, message]
      required: [error]
...
    RegisterRequest:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: "#/components/schemas/UserRole"
      required: [full_name, email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          $ref: "#/components/schemas/User"
      required: [access_token, token_type, user]
    User:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        organization:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, full_name, email, role, created_at, updated_at]
...
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [hackathon, accelerator, incubator, meetup, grant]
        sector:
          type: string
          nullable: true
        topic:
          type: string
          nullable: true
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        location:
          type: string
          nullable: true
        registration_url:
          type: string
          nullable: true
      required: [id, title, category, start_date, end_date]
    EventCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [hackathon, accelerator, incubator, meetup, grant]
        sector:
          type: string
        topic:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        location:
          type: string
        registration_url:
          type: string
          format: uri
      required: [title, category, start_date, end_date]
    EventListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Event"
      required: [items]
...